Backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
httpx
requests
slowapi
groq
email-validator
